src/srv_management/requirements.txt
google-cloud-dns==0.32.3
google-cloud-pubsub==2.6.1
